+++ docs/conf.py
@@ -11,8 +11,8 @@ sys.path.insert(0, str(project_root))
 project = "aiorobokassa"
 copyright = "2025, Masa"
 author = "Masa"
-release = "1.0.1"
-version = "1.0.1"
+release = "1.0.2"
+version = "1.0.2"
 
 # Extensions
 extensions = [